+++ src/client.js
@@ -189,6 +189,117 @@ const clientHandler = (representation, headers) => {
     };
 };
 
+
+
+const fileUploader= (client) => {
+    /**
+     * Will increament the counter
+     * @returns {Promise<number|*>}
+     */
+    const increaseValue= async () => {
+        const soapCall = client._prepareSoapCall('xtk:counter', 'IncreaseValue')
+        soapCall.writeString('name', 'xtkResource')
+        await client._makeSoapCall(soapCall)
+        return soapCall.getNextLong();
+    }
+
+    /**
+     * Will write fileRes
+     * @param counter
+     * @param data
+     * @returns {Promise<*>}
+     */
+    const write= async (counter, data) => {
+        const soapCall = client._prepareSoapCall('xtk:persist', 'Write');
+        const fileRes = soapCall.createElement("fileRes");
+        fileRes.setAttribute("internalName", "RES" + counter);
+        fileRes.setAttribute("md5", data[0].md5);
+        fileRes.setAttribute("label", data[0].fileName);
+        fileRes.setAttribute("fileName", data[0].fileName);
+        fileRes.setAttribute("originalName", data[0].fileName);
+        fileRes.setAttribute("useMd5AsFilename", "1");
+        fileRes.setAttribute("storageType", "5");
+        fileRes.setAttribute("xtkschema", "xtk:fileRes");
+        soapCall.writeDocument("doc", DomUtil.parse(DomUtil.toXMLString(fileRes)));
+        await client._makeSoapCall(soapCall)
+        return fileRes;
+    }
+
+    /**
+     * Will publish the file
+     * @param fileRes
+     * @returns {Promise<void>}
+     */
+    const publishIfNeeded= async(fileRes) => {
+        const soapCall = client._prepareSoapCall('xtk:fileRes', 'PublishIfNeeded');
+        soapCall.writeDocument("entity", DomUtil.parse(DomUtil.toXMLString(fileRes)));
+        //TODO: soapCall.writeDocument("params", new NL.XML.Node("params"));
+        await client._makeSoapCall(soapCall)
+    }
+
+    /**
+     * Will return the public URL of the uploaded file.
+     * @param fileRes
+     * @returns {Promise<string|*>}
+     */
+    const  getPublicUrl = async (fileRes) => {
+        const soapCall = client._prepareSoapCall('xtk:fileRes', 'GetURL');
+        soapCall.writeDocument("entity", DomUtil.parse(DomUtil.toXMLString(fileRes)));
+        await client._makeSoapCall(soapCall)
+        soapCall.getNextElement();
+        return soapCall.getNextString();
+    }
+    return {
+        upload: async (file) => {
+            return new Promise(async (resolve, reject) => {
+                var data = new FormData()
+                data.append('file_noMd5', file)
+                try {
+                    const response = await fetch(`${client._connectionParameters._endpoint}/nl/jsp/uploadFile.jsp`, {
+                        processData: false,
+                        credentials: 'include',
+                        method: 'POST',
+                        body: data,
+                        headers: {
+                            'x-security-token': client._securityToken,
+                            'Cookie': '__sessiontoken=' + client._sessionToken,
+                        }
+                    })
+                    const okay = await response.text()
+                    var iframe = document.createElement('iframe');
+                    iframe.style.height = 0;
+                    iframe.style.width = 0;
+                    document.controller = {
+                        uploadFileCallBack: async (data) => {
+                            const counter = await increaseValue(); // Step 1
+                            const fileRes = await write(counter, data); // Step 2
+                            await publishIfNeeded(fileRes) // Step 3
+                            const url = await getPublicUrl(fileRes) // Step 3
+                            console.log("SUCCESS", url); // Final Output
+                            resolve({
+                                label: data[0].fileName,
+                                md5: data[0].md5,
+                                type: file.type,
+                                size: file.size,
+                                url: url
+                            })
+                        }
+                    }
+                    var html = `<body>${okay}</body>`;
+                    document.body.appendChild(iframe);
+                    iframe.contentWindow.document.open();
+                    iframe.contentWindow.document.write(html);
+                    iframe.contentWindow.document.close();
+                } catch (ex) {
+                    console.log('ec', ex);
+                    reject(ex);
+                }
+            })
+        }
+    }
+}
+
+
 // ========================================================================================
 // Campaign credentials
 // ========================================================================================
@@ -526,7 +637,7 @@ class Client {
         this._refreshClient = connectionParameters._options.refreshClient;
 
         // expose utilities
-        this.fileUploader= new FileUploader(this);
+        this.fileUploader= fileUploader(this);
 
         /**
          * Accessor to DOM helpers
